// indexer/services/scripts/src/helpers/types.ts
import { IndexerTendermintBlock, IndexerTendermintEvent } from '@furyaprotocol-indexer/v4-protos';

export interface AnnotatedIndexerTendermintEvent extends IndexerTendermintEvent {
  data: string;
}

export interface AnnotatedIndexerTendermintBlock extends IndexerTendermintBlock {
  annotatedEvents: AnnotatedIndexerTendermintEvent[];
}

export enum FuryaIndexerSubtypes {
  ORDER_FILL = 'order_fill',
  SUBACCOUNT_UPDATE = 'subaccount_update',
  TRANSFER = 'transfer',
  MARKET = 'market',
  STATEFUL_ORDER = 'stateful_order',
  FUNDING = 'funding_values',
  ASSET = 'asset',
  PERPETUAL_MARKET = 'perpetual_market',
  LIQUIDITY_TIER = 'liquidity_tier',
  UPDATE_PERPETUAL = 'update_perpetual',
  UPDATE_CLOB_PAIR = 'update_clob_pair',
}